pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "navig"
version = "2.4.14"
description = "NAVIG CLI - autonomous infrastructure runtime and control plane"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "NAVIG Development Team", email = "opensource@navig.run" }
]
keywords = ["cli", "devops", "automation", "mcp", "infrastructure", "agent"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]

dependencies = [
  "typer[all]>=0.9.0",
  "rich>=13.7.0",
  "pyyaml>=6.0",
  "requests>=2.31.0",
  "aiohttp>=3.9.0",
  "paramiko>=3.4.0",
  "colorama>=0.4.6",
  "psutil>=5.9.0",
  "pyperclip>=1.8.2",
  "Jinja2>=3.1.0",
  "websockets>=15.0",
  "cryptography>=42.0.0",
  "platformdirs>=4.0.0",
  "httpx>=0.27.0",
  "loguru>=0.7.0",
  "speedtest-cli>=2.1.3",
  "tomli>=2.0; python_version < '3.11'",
  "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "pytest-benchmark>=4.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "build>=1.2.1",
  # Required by memory/embeddings similarity calculation in tests
  "numpy>=1.24.0",
]
voice = [
  "edge-tts>=6.1.9",
  "openai>=1.0.0",
]
keyring = [
  "keyring>=25.0.0",
]
matrix = [
  "matrix-nio[e2e]>=0.24.0",
]
desktop = [
  "uiautomation>=2.0.18",
  "comtypes>=1.2.0",
]
farmore = [
  "farmore>=0.9.0",
]
finance = [
  "beancount>=3.0.0",
]
# telegram extra intentionally empty — NAVIG uses raw Bot API via httpx (no PTB required)
telegram = []
tui = [
  "textual>=0.50.0",
]

[project.urls]
Homepage = "https://github.com/navig-run/core"
Documentation = "https://github.com/navig-run/core/blob/main/docs/README.md"
Repository = "https://github.com/navig-run/core"
Changelog = "https://github.com/navig-run/core/blob/main/CHANGELOG.md"
Security = "https://github.com/navig-run/core/blob/main/SECURITY.md"

[project.scripts]
navig = "navig.main:main"

[tool.setuptools.packages.find]
where = [".", "sdk/python"]
include = ["navig*"]

[tool.setuptools.package-data]
navig = ["*.txt", "*.md", "help/*.md", "schemas/*.json"]
"navig.resources" = ["workflows/*.yaml", "*.md"]
"navig.agent.context" = ["*.md"]
"navig_sdk" = ["*.md"]
# Root content trees included in wheel/sdist
"" = [
    "store/skills/**/*.md",
    "store/skills/**/*.json",
    "store/skills/**/*.txt",
    "store/skills/**/*.yaml",
    "store/skills/**/*.py",
    "store/templates/**/*.yaml",
    "store/templates/**/*.yml",
    "store/templates/**/*.txt",
    "store/formations/**/*.json",
    "store/formations/**/*.md",
    "store/actions/**/*.yaml",
    "store/actions/**/*.json",
    "store/agents/**/*.json",
    "store/agents/**/*.md",
    "store/prompts/**/*.md",
    "store/prompts/**/*.txt",
    "store/stacks/**/*.yaml",
    "store/stacks/**/*.yml",
    "store/workflows/**/*.yaml",
    "store/workflows/**/*.json",
    "packages/**/*.json",
    "packages/**/*.yaml",
    "packages/**/*.yml",
    "packages/**/*.md",
    "packages/**/*.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
minversion = "3.10"
addopts = [
  "-v",
  "--tb=short",
  "--strict-markers",
  "--disable-warnings",
  "--import-mode=importlib",
  "--cov=navig.gateway.server",
  "--cov=navig.daemon.telegram_worker",
  "--cov=navig.daemon.service_manager",
  "--cov=navig.contracts",
  "--cov=navig.remote",
  "--cov=navig.tunnel",
  "--cov=navig.core.security",
  "--cov=navig.routing.router",
  "--cov-report=term-missing",
  "--cov-fail-under=75",
]
asyncio_mode = "auto"
markers = [
  "unit: Unit tests for individual functions",
  "integration: Integration tests requiring mocked dependencies",
  "slow: Tests that take longer to run",
  "requires_server: Tests that require actual server connection",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
